Added my character so far with texture.
I hope it is not bugged as I am in the middle of skinning. I do not know the exact assigned folders. All previous character files can be deleted, but I figured it is best if I do not delete those as I am afraid it will error the heck out of Unity.